Move ENS image fetching out of handler utilities
