+++ src/features/ens/hooks/useENSAvatar.ts
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query';
 
 import { queryClient, type QueryConfigDeprecated, type UseQueryData } from '@/react-query';
 
-import { fetchImage } from '../utils/handlers';
+import { fetchENSImage } from '../utils/fetchENSImage';
 import { getENSData, saveENSData } from '../utils/localStorage';
 
 export const ensAvatarQueryKey = (name: string) => ['ens-avatar', name];
@@ -18,7 +18,7 @@ export async function fetchENSAvatar(name: string, { cacheFirst, swallowError }:
       queryClient.setQueryData(ensAvatarQueryKey(name), cachedAvatar);
       if (cacheFirst) return cachedAvatar as { imageUrl: string };
     }
-    const avatar = await fetchImage('avatar', name);
+    const avatar = await fetchENSImage('avatar', name);
     saveENSData('avatar', name, avatar);
     return avatar;
   } catch (err) {
